[trace-metadata] move callspec code into its own module
Introduce callspec.c/h which encapsulate the call specification
functionality.
