+++ mono/mini/mini.h
@@ -48,6 +48,7 @@
 #include "mono/metadata/marshal.h"
 #include "mono/metadata/security-manager.h"
 #include "mono/metadata/exception.h"
+#include "mono/metadata/callspec.h"
 
 /*
  * The mini code should not have any compile time dependencies on the GC being used, so the same object file from mini/
@@ -537,7 +538,6 @@ typedef struct MonoMethodVar MonoMethodVar;
 typedef struct MonoBasicBlock MonoBasicBlock;
 typedef struct MonoLMF MonoLMF;
 typedef struct MonoSpillInfo MonoSpillInfo;
-typedef struct MonoCallSpec MonoCallSpec;
 
 extern MonoCallSpec *mono_jit_trace_calls;
 extern gboolean mono_break_on_exc;
@@ -2992,15 +2992,6 @@ MONO_API void      mono_debugger_run_finally             (MonoContext *start_ctx
 MONO_API gboolean mono_breakpoint_clean_code (guint8 *method_start, guint8 *code, int offset, guint8 *buf, int size);
 
 /* Tracing */
-MONO_PROFILER_API gboolean mono_callspec_parse		(const char *options,
-							 MonoCallSpec *spec);
-MONO_PROFILER_API void     mono_callspec_cleanup	(MonoCallSpec *spec);
-MONO_PROFILER_API gboolean mono_callspec_eval_exception	(MonoClass *klass,
-							 MonoCallSpec *spec);
-MONO_PROFILER_API gboolean mono_callspec_eval		(MonoMethod *method,
-							 const MonoCallSpec *spec);
-void			   mono_callspec_set_assembly	(MonoAssembly *assembly);
-
 MonoCallSpec *mono_trace_set_options           (const char *options);
 gboolean       mono_trace_eval                  (MonoMethod *method);
