+++ mono/mini/trace.h
@@ -7,29 +7,6 @@
 #include <glib.h>
 #include "mono/utils/mono-compiler.h"
 
-typedef enum {
-	MONO_TRACEOP_ALL,
-	MONO_TRACEOP_PROGRAM,
-	MONO_TRACEOP_METHOD,
-	MONO_TRACEOP_ASSEMBLY,
-	MONO_TRACEOP_CLASS,
-	MONO_TRACEOP_NAMESPACE,
-	MONO_TRACEOP_EXCEPTION,
-	MONO_TRACEOP_WRAPPER,
-} MonoTraceOpcode;
-
-typedef struct {
-	MonoTraceOpcode op;
-	int   exclude;
-	void *data, *data2;
-} MonoTraceOperation;
-
-struct MonoCallSpec {
-	int len;
-	gboolean enabled;
-	MonoTraceOperation *ops;
-};
-
 G_BEGIN_DECLS
 
 void
